+++ pkg/backend/filestate/backend.go
@@ -348,8 +348,9 @@ func (b *localBackend) apply(
 
 	// Print a banner so it's clear this is a local deployment.
 	actionLabel := backend.ActionLabel(kind, opts.DryRun)
-	fmt.Printf(op.Opts.Display.Color.Colorize(
+	_, err := fmt.Fprintf(os.Stderr, op.Opts.Display.Color.Colorize(
 		colors.SpecHeadline+"%s (%s):"+colors.Reset+"\n"), actionLabel, stackRef)
+	contract.IgnoreError(err)
 
 	// Start the update.
 	update, err := b.newUpdate(stackName, op.Proj, op.Root)
