+++ tests/integration/integration_test.go
@@ -8,6 +8,7 @@ import (
 	"os"
 	"path"
 	"path/filepath"
+	"regexp"
 	"strings"
 	"testing"
 	"time"
@@ -127,8 +128,8 @@ func TestProjectMain(t *testing.T) {
 		e.ImportDirectory("project_main_abs")
 		e.RunCommand("pulumi", "login", "--cloud-url", e.LocalURL())
 		e.RunCommand("pulumi", "stack", "init", "main-abs")
-		stdout, stderr := e.RunCommandExpectError("pulumi", "up", "--non-interactive", "--skip-preview")
-		assert.Equal(t, "Updating (main-abs):\n", stdout)
+		_, stderr := e.RunCommandExpectError("pulumi", "up", "--non-interactive", "--skip-preview")
+		assert.Contains(t, stderr, "Updating (main-abs):")
 		assert.Contains(t, stderr, "project 'main' must be a relative path")
 		e.RunCommand("pulumi", "stack", "rm", "--yes")
 	})
@@ -143,8 +144,8 @@ func TestProjectMain(t *testing.T) {
 		e.ImportDirectory("project_main_parent")
 		e.RunCommand("pulumi", "login", "--cloud-url", e.LocalURL())
 		e.RunCommand("pulumi", "stack", "init", "main-parent")
-		stdout, stderr := e.RunCommandExpectError("pulumi", "up", "--non-interactive", "--skip-preview")
-		assert.Equal(t, "Updating (main-parent):\n", stdout)
+		_, stderr := e.RunCommandExpectError("pulumi", "up", "--non-interactive", "--skip-preview")
+		assert.Contains(t, stderr, "Updating (main-parent):")
 		assert.Contains(t, stderr, "project 'main' must be a subfolder")
 		e.RunCommand("pulumi", "stack", "rm", "--yes")
 	})
@@ -231,6 +232,73 @@ func TestRemoveWithResourcesBlocked(t *testing.T) {
 	_, stderr := e.RunCommandExpectError("pulumi", "stack", "rm", "--yes")
 	assert.Contains(t, stderr, "--force")
 	e.RunCommand("pulumi", "destroy", "--skip-preview", "--non-interactive", "--yes")
+	e.RunCommand("pulumi", "stack", "rm", "--yes")
+}
+
+// TestPreviewJSON tests serializing a preview as JSON using the `pulumi preview --json` option.
+func TestPreviewJSON(t *testing.T) {
+	if os.Getenv("PULUMI_ACCESS_TOKEN") == "" {
+		t.Skipf("Skipping: PULUMI_ACCESS_TOKEN is not set")
+	}
+
+	e := ptesting.NewEnvironment(t)
+	defer func() {
+		if !t.Failed() {
+			e.DeleteEnvironment()
+		}
+	}()
+
+	stackName, err := resource.NewUniqueHex("rm-test-", 8, -1)
+	contract.AssertNoErrorf(err, "resource.NewUniqueHex sould not fail with no maximum length is set")
+
+	e.ImportDirectory("single_resource")
+	e.RunCommand("pulumi", "stack", "init", stackName)
+	e.RunCommand("yarn", "link", "@pulumi/pulumi")
+
+	// Run the preview to get the JSON. Trim out some pieces we don't want to be sensitive to (like
+	// the precise details of dynamic provider serialization), and then ensure the text matches what we expect.
+	stdout, _ := e.RunCommand("pulumi", "preview", "--json")
+	// nolint: lll
+	expect := fmt.Sprintf(`{
+    "steps": [
+        {
+            "op": "create",
+            "urn": "urn:pulumi:%[1]s::protect_resources::pulumi:pulumi:Stack::protect_resources-%[1]s",
+            "newState": {
+                "urn": "urn:pulumi:%[1]s::protect_resources::pulumi:pulumi:Stack::protect_resources-%[1]s",
+                "custom": false,
+                "type": "pulumi:pulumi:Stack"
+            }
+        },
+        {
+            "op": "create",
+            "urn": "urn:pulumi:%[1]s::protect_resources::pulumi-nodejs:dynamic:Resource::res",
+            "provider": "urn:pulumi:%[1]s::protect_resources::pulumi:providers:pulumi-nodejs::default::04da6b54-80e4-46f7-96ec-b56ff0331ba9",
+            "newState": {
+                "urn": "urn:pulumi:%[1]s::protect_resources::pulumi-nodejs:dynamic:Resource::res",
+                "custom": true,
+                "type": "pulumi-nodejs:dynamic:Resource",
+                "inputs": {
+                    "__provider": "...",
+                    "state": 1
+                },
+                "parent": "urn:pulumi:%[1]s::protect_resources::pulumi:pulumi:Stack::protect_resources-%[1]s",
+                "provider": "urn:pulumi:%[1]s::protect_resources::pulumi:providers:pulumi-nodejs::default::04da6b54-80e4-46f7-96ec-b56ff0331ba9",
+                "propertyDependencies": {
+                    "__provider": null,
+                    "state": null
+                }
+            }
+        }
+    ],
+    "changeSummary": {
+        "create": 2
+    }
+}
+`, stackName)
+	rex := regexp.MustCompile(`"__provider": ".*",`)
+	stdout = rex.ReplaceAllString(stdout, `"__provider": "...",`)
+	assert.Equal(t, expect, stdout)
 	e.RunCommand("pulumi", "stack", "rm", "--yes")
 }
